Currency entity fix
