+++ TglCA/TglCA.Dal.Interfaces/Entities/Currency.cs
@@ -1,16 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TglCA.Dal.Interfaces.Entities.Base;
+﻿using TglCA.Dal.Interfaces.Entities.Base;
 
 namespace TglCA.Dal.Interfaces.Entities
 {
     public class Currency : BaseEntity
     {
-        public byte[] Img { get; set; }
+        public string CurrencyId { get; set; }
         public string Name { get; set; }
         public string Symbol { get; set; }
+        public bool IsDeleted { get; set; } = false;
     }
 }
